requirements.txt
alembic==1.15.2
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
Flask==3.1.0
flask-cors==5.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
git-filter-repo==2.47.0
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
packaging==25.0
pillow==11.2.1
python-dotenv==1.1.0
qrcode==8.2
requests==2.32.3
SQLAlchemy==2.0.40
stripe==12.1.0
typing_extensions==4.13.0
urllib3==2.4.0
Werkzeug==3.1.3
